requirements.txt
asgiref==3.5.2
autopep8==1.7.0
beautifulsoup4==4.11.1
blinker==1.5
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cssselect==1.1.0
Django==4.1.2
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-js-asset==2.0.0
djangorestframework==3.14.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.8.0
greenlet==1.1.3.post0
gunicorn==20.0.4
idna==3.4
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.1
MarkupSafe==2.1.1
mysql==0.0.3
mysql-connector-python==8.0.31
mysqlclient==2.1.1
newspaper3k==0.2.8
nltk==3.7
numpy==1.23.4
Pillow==9.2.0
protobuf==3.20.1
pycodestyle==2.9.1
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tinysegmenter==0.3
tldextract==3.4.0
toml==0.10.2
tqdm==4.64.1
tzdata==2022.5
urllib3==1.26.12
Werkzeug==2.2.2
whitenoise==6.3.0

